+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -459,6 +459,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed some typos in debug menu (case 1224594)
 - Fixed ray traced point and spot lights shadows not rejecting istory when semi-transparent or colored.
 - Fixed a warning due to StaticLightingSky when reloading domain in some cases.
+- Fixed the MaxLightCount being displayed when the light volume debug menu is on ColorAndEdge.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled